refactor: extract client-side auth into auth-context module
